Added smart loading for the application
Instead of loading the application ahead of time, and not allowing the user to go to the next view, or loading it when they need it, making them wait, I found a way to load ahead of time from the menu controller, and then update the UI asnchrnously if it hasn't loaded, or instantly update the view if it has!

This requires making a reference to the child/future controller, but it appears to work well.